+++ src/npctalk.cpp
@@ -1520,7 +1520,7 @@ talk_topic dialogue::opt( dialogue_window &d_win, const talk_topic &topic )
         response_lines.push_back( responses[i].create_option_line( *this, 'a' + i ) );
     }
 
-    long ch = text_only ? 'a' + responses.size() - 1 : ' ';
+    int ch = text_only ? 'a' + responses.size() - 1 : ' ';
     bool okay;
     do {
         d_win.refresh_response_display();
@@ -1544,7 +1544,7 @@ talk_topic dialogue::opt( dialogue_window &d_win, const talk_topic &topic )
                     ch -= 'a';
                     break;
             }
-        } while( ( ch < 0 || ch >= static_cast<long>( responses.size() ) ) );
+        } while( ( ch < 0 || ch >= static_cast<int>( responses.size() ) ) );
         okay = true;
         std::set<dialogue_consequence> consequences = responses[ch].get_consequences( *this );
         if( consequences.count( dialogue_consequence::hostile ) > 0 ) {
@@ -2784,7 +2784,7 @@ void conditional_t::set_npc_allies( JsonObject &jo )
 
 void conditional_t::set_npc_service( JsonObject &jo )
 {
-    const unsigned long service_price = jo.get_int( "npc_service" );
+    const signed long service_price = jo.get_int( "npc_service" );
     condition = [service_price]( const dialogue & d ) {
         return !d.beta->has_effect( effect_currently_busy ) && d.alpha->cash >= service_price;
     };
@@ -2792,7 +2792,7 @@ void conditional_t::set_npc_service( JsonObject &jo )
 
 void conditional_t::set_u_has_cash( JsonObject &jo )
 {
-    const unsigned long min_cash = jo.get_int( "u_has_cash" );
+    const signed long min_cash = jo.get_int( "u_has_cash" );
     condition = [min_cash]( const dialogue & d ) {
         return d.alpha->cash >= min_cash;
     };
